+++ projects/core/src/data-providers/sql-database.ts
@@ -62,10 +62,11 @@ import { isOfType } from '../isOfType.js'
  */
 export class SqlDatabase
   implements
-  DataProvider,
-  HasWrapIdentifier,
-  CanBuildMigrations,
-  SqlCommandFactory {
+    DataProvider,
+    HasWrapIdentifier,
+    CanBuildMigrations,
+    SqlCommandFactory
+{
   /**
    * Gets the SQL database from the data provider.
    * @param dataProvider - The data provider.
@@ -223,7 +224,7 @@ export class SqlDatabase
     var b = new FilterConsumerBridgeToSqlRequest(
       sqlCommand,
       dbNames ||
-      (await dbNamesOfWithForceSqlExpression(r.metadata, wrapIdentifier)),
+        (await dbNamesOfWithForceSqlExpression(r.metadata, wrapIdentifier)),
     )
     b._addWhere = false
     await (
@@ -289,7 +290,7 @@ class LogSQLCommand implements SqlCommand {
   constructor(
     private origin: SqlCommand,
     private logToConsole: typeof SqlDatabase.LogToConsole,
-  ) { }
+  ) {}
 
   args: any = {}
   addParameterAndReturnSqlToken(val: any) {
@@ -344,7 +345,7 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
     private sql: SqlDatabase,
     private iAmUsed: (e: EntityDbNamesBase) => Promise<void>,
     private strategy: SqlImplementation,
-  ) { }
+  ) {}
 
   async init() {
     let dbNameProvider: EntityDbNamesBase =
@@ -383,7 +384,7 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
   async find(options?: EntityDataProviderFindOptions): Promise<any[]> {
     let e = await this.init()
 
-    let { colKeys, select } = this.buildSelect(e)
+    let { colKeys, select } = this.buildSelect(e, options?.select)
     select = 'select ' + select
 
     select += '\n from ' + e.$entityName
@@ -456,10 +457,11 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
     return result
   }
 
-  private buildSelect(e: EntityDbNamesBase) {
+  private buildSelect(e: EntityDbNamesBase, selectedFields?: string[]) {
     let select = ''
     let colKeys: FieldMetadata[] = []
     for (const x of this.entity.fields) {
+      if (selectedFields && !selectedFields.includes(x.key)) continue
       if (x.isServerExpression) {
       } else {
         if (colKeys.length > 0) select += ', '
@@ -514,9 +516,9 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
       if (sqlResult.rows.length != 1)
         throw new Error(
           'Failed to update row with id ' +
-          id +
-          ', rows updated: ' +
-          sqlResult.rows.length,
+            id +
+            ', rows updated: ' +
+            sqlResult.rows.length,
         )
       return this.buildResultRow(colKeys, sqlResult.rows[0], sqlResult)
     })
@@ -574,7 +576,7 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
           if (typeof id !== 'number')
             throw new Error(
               'Auto increment, for a database that is does not support returning syntax, should return an array with the single last added id. Instead it returned: ' +
-              JSON.stringify(id),
+                JSON.stringify(id),
             )
           return this.find({
             where: new Filter((x) =>
@@ -669,8 +671,9 @@ export async function groupByImpl(
         if (x.isServerExpression) {
         } else {
           const dbName = await e.$dbNameOf(x)
-          select += `, ${aggregateSqlSyntax(operator, dbName)} as ${x.key
-            }_${operator}`
+          select += `, ${aggregateSqlSyntax(operator, dbName)} as ${
+            x.key
+          }_${operator}`
         }
 
         const turnToNumber =